refactor: remove useConversation hook and update useSearch to utilize useNewConvo
