+++ client/src/hooks/Conversations/useSearch.ts
@@ -5,14 +5,23 @@ import { useGetSearchEnabledQuery } from 'librechat-data-provider/react-query';
 import type { UseInfiniteQueryResult } from '@tanstack/react-query';
 import type { ConversationListResponse } from 'librechat-data-provider';
 import { useSearchInfiniteQuery } from '~/data-provider';
-import useConversation from './useConversation';
+import useNewConvo from '~/hooks/useNewConvo';
 import store from '~/store';
 
 export default function useSearchMessages({ isAuthenticated }: { isAuthenticated: boolean }) {
   const navigate = useNavigate();
   const location = useLocation();
   const [pageNumber, setPageNumber] = useState(1);
-  const { searchPlaceholderConversation } = useConversation();
+  const { switchToConversation } = useNewConvo();
+  const searchPlaceholderConversation = useCallback(() => {
+    switchToConversation({
+      conversationId: 'search',
+      title: 'Search',
+      endpoint: null,
+      createdAt: '',
+      updatedAt: '',
+    });
+  }, [switchToConversation]);
 
   const searchQuery = useRecoilValue(store.searchQuery);
   const setIsSearchEnabled = useSetRecoilState(store.isSearchEnabled);
